Egghead.MongoDbStorage: All stores interfaces are moved to storage layer
Switched back props types from double to long: likes, dislikes, e.t.c.
All of results from storage are returned to caller as a mongo objects. (In the near future the intermediate objects will be created)

+++ Egghead.MongoDbStorage/Articles/MongoDbArticle.cs
@@ -18,10 +18,10 @@ namespace Egghead.MongoDbStorage.Articles
         public string Text { get; set; }      
         public string NormalizedEmail { get; set; }
         public ReleaseType ReleaseType { get; set; }
-        public double LikesCount { get; set; }
-        public double DislikesCount { get; set; }
-        public double ViewsCount { get; set; }
-        public double CommentsCount { get; set; }
+        public long LikesCount { get; set; }
+        public long DislikesCount { get; set; }
+        public long ViewsCount { get; set; }
+        public long CommentsCount { get; set; }
         public DateTime CreatedAt { get; set; }
         public DateTime ChangedAt { get; set; }
         public DateTime DeletedAt { get; set; }
